Move async barrier to util
